+++ src/main/scala/apg/Q.scala
@@ -6,11 +6,11 @@ object Q {
 
   NativeUtils.loadLibraryFromJar("/apg.jni")
 
-  @native def findOrthogonalVector(N: Int, u: Double, v: Double, gamma: Double): Array[Double]
+  @native def findOrthogonalVector(N: Int, u: Double, v: Double, gamma: Double): Array[Float]
 
   def expQTtx(degree: Int, steps: Int, N: Int, u: Double, v: Double, gamma: Double, t: Double, x: F): F =
     new F(N, expQTtx(degree, steps, N, u, v, gamma, t, x.f))
 
-  @native private[this] def expQTtx(degree: Int, steps: Int, N: Int, u: Double, v: Double, gamma: Double, t: Double, x: Array[Double]): Array[Double]
+  @native private[this] def expQTtx(degree: Int, steps: Int, N: Int, u: Double, v: Double, gamma: Double, t: Double, x: Array[Float]): Array[Float]
 
 }
